feat: Enhance testing infrastructure by updating UBSan test exclusions and improving memory initialization in tests
