+++ OloEngine/src/OloEngine/Memory/TypeTraits.h
@@ -648,30 +648,6 @@ namespace OloEngine
     template<typename T>
     inline constexpr bool TIsTriviallyRelocatable_V = TIsTriviallyRelocatable<T>::Value;
 
-    // std::basic_string uses SSO with self-referencing pointers on some implementations
-    // (notably libstdc++), making it unsafe to relocate via memmove.
-    template<typename CharT, typename Traits, typename Alloc>
-    struct TIsTriviallyRelocatable<std::basic_string<CharT, Traits, Alloc>>
-    {
-        enum
-        {
-            Value = false
-        };
-    };
-
-    // TTuple (and TPair alias) is only trivially relocatable if ALL its element types are.
-    template<typename... Types>
-    struct TTuple;
-
-    template<typename... Types>
-    struct TIsTriviallyRelocatable<TTuple<Types...>>
-    {
-        enum
-        {
-            Value = (TIsTriviallyRelocatable<Types>::Value && ...)
-        };
-    };
-
     // @struct TUseBitwiseSwap
     // @brief Determines if bitwise operations (memcpy/memswap) should be used for relocation
     //
